refactor(test): fixture를 도메인 모듈 testFixtures로 이동, ExceptionAssertions :test-support로 통합
- ADR-0015: PlayerFixture/RoomFixture 등 domain fixture를 각 모듈 src/testFixtures/로 이동
  (room·game·user·admin). app:test의 중복 복사본 제거
- ADR-0015: ExceptionAssertions 6개 모듈 로컬 복사본 제거, coffeeshout.support 단일 버전으로 통합.
  admin·game·user·zzolbot build.gradle.kts에 testImplementation(project(":test-support")) 추가
