+++ backend/admin/build.gradle.kts
@@ -1,5 +1,7 @@
 // :admin — 운영자 영역 (dashboard + patchnote + report)
 
+apply(plugin = "java-test-fixtures")
+
 tasks.bootJar { enabled = false }
 tasks.jar { enabled = true }
 
@@ -32,4 +34,8 @@ dependencies {
     annotationProcessor("com.querydsl:querydsl-apt:$queryDslVersion:jakarta")
     annotationProcessor("jakarta.annotation:jakarta.annotation-api")
     annotationProcessor("jakarta.persistence:jakarta.persistence-api")
+
+    "testFixturesImplementation"(project(":game-api"))
+
+    testImplementation(project(":test-support"))
 }
