+++ backend/admin/src/test/java/coffeeshout/patchnote/infra/persistence/PatchNoteEntityTest.java
@@ -1,6 +1,6 @@
 package coffeeshout.patchnote.infra.persistence;
 
-import static coffeeshout.fixture.ExceptionAssertions.assertCoffeeShoutException;
+import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
