+++ backend/game/build.gradle.kts
@@ -1,5 +1,7 @@
 // :game — 6게임 + orchestration (이전 minigame)
 
+apply(plugin = "java-test-fixtures")
+
 tasks.bootJar { enabled = false }
 tasks.jar { enabled = true }
 
@@ -21,4 +23,10 @@ dependencies {
     annotationProcessor("com.querydsl:querydsl-apt:$queryDslVersion:jakarta")
     annotationProcessor("jakarta.annotation:jakarta.annotation-api")
     annotationProcessor("jakarta.persistence:jakarta.persistence-api")
+
+    "testFixturesImplementation"(project(":game-api"))
+    "testFixturesImplementation"(testFixtures(project(":room")))
+
+    testImplementation(testFixtures(project(":room")))
+    testImplementation(project(":test-support"))
 }
