+++ backend/game/src/test/java/coffeeshout/blockstacking/domain/BlockStackingGameTest.java
@@ -1,6 +1,6 @@
 package coffeeshout.blockstacking.domain;
 
-import static coffeeshout.fixture.ExceptionAssertions.assertCoffeeShoutException;
+import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import coffeeshout.fixture.PlayerFixture;
