+++ backend/game/src/test/java/coffeeshout/laddergame/domain/LadderGameTest.java
@@ -1,6 +1,6 @@
 package coffeeshout.laddergame.domain;
 
-import static coffeeshout.fixture.ExceptionAssertions.assertCoffeeShoutException;
+import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import coffeeshout.fixture.PlayerFixture;
