+++ backend/game/src/testFixtures/java/coffeeshout/fixture/CardGameDeckStub.java
