+++ backend/game/src/testFixtures/java/coffeeshout/fixture/MiniGameDummy.java
