+++ backend/room/src/testFixtures/java/coffeeshout/fixture/FixedLastValueGenerator.java
