+++ backend/room/src/testFixtures/java/coffeeshout/fixture/RoomSessionClaimFixture.java
@@ -2,7 +2,10 @@ package coffeeshout.fixture;
 
 import coffeeshout.room.infra.auth.RoomSessionClaim;
 
-public class RoomSessionClaimFixture {
+public final class RoomSessionClaimFixture {
+
+    private RoomSessionClaimFixture() {
+    }
 
     public static RoomSessionClaim 로그인_플레이어() {
         return RoomSessionClaim.of("ABCD", "홍길동", 42L);
