+++ backend/room/src/testFixtures/java/coffeeshout/fixture/RouletteFixture.java
