+++ backend/user/src/test/java/coffeeshout/friend/domain/FriendshipTest.java
@@ -1,6 +1,6 @@
 package coffeeshout.friend.domain;
 
-import static coffeeshout.fixture.ExceptionAssertions.assertCoffeeShoutException;
+import static coffeeshout.support.ExceptionAssertions.assertCoffeeShoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
