+++ backend/zzolbot/build.gradle.kts
@@ -21,4 +21,5 @@ dependencies {
     implementation("com.github.jsqlparser:jsqlparser:5.0")
     implementation("io.github.resilience4j:resilience4j-spring-boot3:$resilience4jVersion")
 
+    testImplementation(project(":test-support"))
 }
